Move TensorWrappedPhiTensorPointer to itse own file
